+++ src/iTextSharp.LGPLv2.Core/iTextSharp/text/pdf/TrueTypeFont.cs
@@ -438,7 +438,7 @@ internal class TrueTypeFont : BaseFont
             case AWT_MAXADVANCE:
                 return fontSize * Hhea.AdvanceWidthMax / Head.UnitsPerEm;
             case UNDERLINE_POSITION:
-                return (UnderlinePosition - UnderlineThickness / 2) * fontSize / Head.UnitsPerEm;
+                return (UnderlinePosition - (float)UnderlineThickness / 2) * fontSize / Head.UnitsPerEm;
             case UNDERLINE_THICKNESS:
                 return UnderlineThickness * fontSize / Head.UnitsPerEm;
             case STRIKETHROUGH_POSITION:
@@ -1301,8 +1301,12 @@ internal class TrueTypeFont : BaseFont
                 byte[] b = null;
                 if (subsetp || DirectoryOffset != 0 || SubsetRanges != null)
                 {
-                    var sb = new TrueTypeFontSubSet(FileName, new RandomAccessFileOrArray(Rf), glyphs, DirectoryOffset,
-                                                    true, !subsetp);
+                    var sb = new TrueTypeFontSubSet(FileName,
+                                                    new RandomAccessFileOrArray(Rf),
+                                                    glyphs,
+                                                    DirectoryOffset,
+                                                    true,
+                                                    !subsetp);
                     b = sb.Process();
                 }
                 else
@@ -1478,8 +1482,11 @@ internal class TrueTypeFont : BaseFont
     /// <param name="shortTag">a 256 bytes long  byte  array where each unused byte is represented by 0</param>
     /// <param name="fontDescriptor">the indirect reference to a PdfDictionary containing the font descriptor or  null </param>
     /// <returns>the PdfDictionary containing the font dictionary</returns>
-    protected PdfDictionary GetFontBaseType(PdfIndirectReference fontDescriptor, string subsetPrefix, int firstChar,
-                                            int lastChar, byte[] shortTag)
+    protected PdfDictionary GetFontBaseType(PdfIndirectReference fontDescriptor,
+                                            string subsetPrefix,
+                                            int firstChar,
+                                            int lastChar,
+                                            byte[] shortTag)
     {
         var dic = new PdfDictionary(PdfName.Font);
         if (Cff)
@@ -1573,18 +1580,19 @@ internal class TrueTypeFont : BaseFont
     /// <param name="fontStream">the indirect reference to a PdfStream containing the font or  null </param>
     /// <param name="cidset"></param>
     /// <returns>the PdfDictionary containing the font descriptor or  null </returns>
-    protected PdfDictionary GetFontDescriptor(PdfIndirectReference fontStream, string subsetPrefix,
+    protected PdfDictionary GetFontDescriptor(PdfIndirectReference fontStream,
+                                              string subsetPrefix,
                                               PdfIndirectReference cidset)
     {
         var dic = new PdfDictionary(PdfName.Fontdescriptor);
         dic.Put(PdfName.Ascent, new PdfNumber(Os2.STypoAscender * 1000 / Head.UnitsPerEm));
         dic.Put(PdfName.Capheight, new PdfNumber(Os2.SCapHeight * 1000 / Head.UnitsPerEm));
         dic.Put(PdfName.Descent, new PdfNumber(Os2.STypoDescender * 1000 / Head.UnitsPerEm));
-        dic.Put(PdfName.Fontbbox, new PdfRectangle(
-                                                   Head.XMin * 1000 / Head.UnitsPerEm,
-                                                   Head.YMin * 1000 / Head.UnitsPerEm,
-                                                   Head.XMax * 1000 / Head.UnitsPerEm,
-                                                   Head.YMax * 1000 / Head.UnitsPerEm));
+        dic.Put(PdfName.Fontbbox,
+                new PdfRectangle((float)Head.XMin * 1000 / Head.UnitsPerEm,
+                                 (float)Head.YMin * 1000 / Head.UnitsPerEm,
+                                 (float)Head.XMax * 1000 / Head.UnitsPerEm,
+                                 (float)Head.YMax * 1000 / Head.UnitsPerEm));
         if (cidset != null)
         {
             dic.Put(PdfName.Cidset, cidset);
